requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-allauth==0.63.6
django-environ==0.11.2
django-oauth-toolkit==2.4.0
django-rest-framework-social-oauth2==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gunicorn==20.0.4
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
